refactor: 최소시간 또는 최단거리 그래프를 조회요청을 받은 후에 생성하도록 변경
