+++ src/index.js
@@ -1,12 +1,8 @@
-import {
-  loadSubwayMapDataToLocalStorage,
-  makeSubwayMapGraphAll,
-} from './managers/data-storage-manager.js';
+import { loadSubwayMapDataToLocalStorage } from './managers/data-storage-manager.js';
 import { appendView } from './managers/view-manager.js';
 
 export default function SubwayPathFinder() {
   loadSubwayMapDataToLocalStorage();
-  makeSubwayMapGraphAll();
   appendView();
 }
